Update the names of the telemetry providers so there is no duplication (#59)
* Add AudioClassDriver to name of A2dp manifested provider

The manifested and telemetry providers had the same name. This change renames the manifested provider so that both names follow naming conventions but are distinct fro each other.

* Update BluetoothStack.wprp

* Update BluetoothStack.wprp

* Split A2dp and add base provider

This change splits the A2dp provider into two separate levels for the verbose and light profiles.
It also adds the base telemetry provider to both the verbose and light profiles.